+++ atlastk/XDHq.py
@@ -1,3 +1,5 @@
+# coding: utf-8
+
 """
 MIT License
 
@@ -25,6 +27,7 @@ SOFTWARE.
 import XDHqDEMO, XDHqSHRD, XDHqXML
 
 import os, sys
+from collections import OrderedDict
 
 if sys.version_info[0] == 2:
 	import urllib
@@ -57,7 +60,10 @@ def _split(keysAndValues):
 
 def _unsplit(keys,values):
 	i = 0
-	keysAndValues = {}
+	# With 'OrderedDict', the order of the items is keeped under Python 2.
+	# This facilitates the retrieving of values by using 'values()' method.
+	# Dict are ordered by default under Python 3.
+	keysAndValues = OrderedDict()
 	length = len(keys)
 
 	while i < length: